Test checkpoints without offering metadata
